requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
email-validator==1.1.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Pillow==5.4.1
requests==2.21.0
RPi.GPIO==0.7.0
secrets==1.0.2
SQLAlchemy==1.3.18
WTForms==2.3.3